requirements.txt
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.2
scipy==1.10.1
seaborn==0.12.2
statsmodels==0.14.0
openpyxl==3.1.2
pyyaml==6.0.1 
black==23.3.0
autoflake==2.3.1
scikit-learn==1.5.2
torch>=2.0.0
